+++ ichnaea/api/locate/query.py
@@ -2,6 +2,15 @@
 Code representing a location query.
 """
 
+import ipaddr
+
+from ichnaea.api.locate.constants import MIN_WIFIS_IN_QUERY
+from ichnaea.api.locate.schema import (
+    CellAreaLookup,
+    CellLookup,
+    WifiLookup,
+)
+
 
 class Query(object):
 
@@ -21,15 +30,78 @@ class Query(object):
         :param fallbacks: A dictionary of fallback options.
         :type fallbacks: dict
         """
-        if not geoip:
-            geoip = None
         self.geoip = geoip
-        if not cell:
-            cell = []
         self.cell = cell
-        if not wifi:
-            wifi = []
         self.wifi = wifi
-        if not fallbacks:
-            fallbacks = {}
         self.fallbacks = fallbacks
+
+    @property
+    def geoip(self):
+        return self._geoip
+
+    @geoip.setter
+    def geoip(self, value):
+        if not value:
+            value = None
+        try:
+            valid = str(ipaddr.IPAddress(value))
+        except ValueError:
+            valid = None
+        self._geoip = valid
+
+    @property
+    def cell(self):
+        return self._cell
+
+    @property
+    def cell_area(self):
+        return self._cell_area
+
+    @cell.setter
+    def cell(self, values):
+        if not values:
+            values = []
+        values = list(values)
+        self._cell_unvalidated = values
+
+        filtered_areas = []
+        filtered_cells = []
+        for value in values:
+            valid_area = CellAreaLookup.validate(value)
+            if valid_area:
+                filtered_areas.append(valid_area)
+            valid_cell = CellLookup.validate(value)
+            if valid_cell:
+                filtered_cells.append(valid_cell)
+        self._cell_area = filtered_areas
+        self._cell = filtered_cells
+
+    @property
+    def wifi(self):
+        return self._wifi
+
+    @wifi.setter
+    def wifi(self, values):
+        if not values:
+            values = []
+        values = list(values)
+        self._wifi_unvalidated = values
+
+        filtered = []
+        for value in values:
+            valid = WifiLookup.validate(value)
+            if valid:
+                filtered.append(valid)
+        if len(filtered) < MIN_WIFIS_IN_QUERY:
+            filtered = []
+        self._wifi = filtered
+
+    @property
+    def fallbacks(self):
+        return self._fallbacks
+
+    @fallbacks.setter
+    def fallbacks(self, values):
+        if not values:
+            values = {}
+        self._fallbacks = values